refactor(cli): centralize command help
